chore: move domain test files to their respective domain folders
