+++ src/server/fetcher/fetcher.test.ts
@@ -1,12 +1,12 @@
 import * as oauth from "oauth4webapi";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
-import { getDefaultRoutes } from "../test-fixtures/defaults.js";
-import { generateSecret } from "../test-fixtures/utils.js";
-import { AuthClient } from "./auth-client.js";
-import { Fetcher } from "./fetcher.js";
-import { StatelessSessionStore } from "./session/stateless-session-store.js";
-import { TransactionStore } from "./transaction-store.js";
+import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
+import { generateSecret } from "../../test-fixtures/utils.js";
+import { AuthClient } from "../auth-client.js";
+import { Fetcher } from "../fetcher.js";
+import { StatelessSessionStore } from "../session/stateless-session-store.js";
+import { TransactionStore } from "../transaction-store.js";
 
 // Mock oauth4webapi
 vi.mock("oauth4webapi", async () => {
