+++ src/server/mfa/mfa-popup.test.ts
@@ -3,15 +3,15 @@ import * as jose from "jose";
 import * as oauth from "oauth4webapi";
 import { describe, expect, it, vi } from "vitest";
 
-import { getDefaultRoutes } from "../test-fixtures/defaults.js";
-import { generateSecret, stripTransactionValuePrefix } from "../test-fixtures/utils.js";
-import { RESPONSE_TYPES, SessionData } from "../types/index.js";
-import { createAuthCompletePostMessageResponse } from "../utils/html-helpers.js";
-import { AuthClient } from "./auth-client.js";
-import { decrypt, encrypt } from "./cookies.js";
-import { DiscoveryCache } from "./discovery-cache.js";
-import { StatelessSessionStore } from "./session/stateless-session-store.js";
-import { TransactionState, TransactionStore } from "./transaction-store.js";
+import { getDefaultRoutes } from "../../test-fixtures/defaults.js";
+import { generateSecret, stripTransactionValuePrefix } from "../../test-fixtures/utils.js";
+import { RESPONSE_TYPES, SessionData } from "../../types/index.js";
+import { createAuthCompletePostMessageResponse } from "../../utils/html-helpers.js";
+import { AuthClient } from "../auth-client.js";
+import { decrypt, encrypt } from "../cookies.js";
+import { DiscoveryCache } from "../discovery-cache.js";
+import { StatelessSessionStore } from "../session/stateless-session-store.js";
+import { TransactionState, TransactionStore } from "../transaction-store.js";
 
 function createSessionData(sessionData: Partial<SessionData>): SessionData {
   return {
